+++ CHANGELOG.md
@@ -9,6 +9,8 @@ This is a maintenance release:
    will be used for subsequent (new) connections to a broker.
  * Channel based producer (Producer `ProduceChannel()`) and channel based
    consumer (Consumer `Events()`) are deprecated.
+ * Added `IsTimeout()` on Error type. This is a convenience method that checks
+   if the error is due to a timeout.
 
 
 ## v1.9.2